// src/lib/pg/repository/types.ts
export type TCompareQueryResult = { query: string; values?: unknown[]; };
export type TConditionalDomainFieldsType<First, Second> = First extends TDomainFields ? First : Partial<Second>;
export type TConditionalRawParamsType<First, Second> = First extends SharedTypes.TRawParams ? First : Partial<Second>;
export type TDBCreds = pg.PoolConfig & { database: string; host: string; password: string; port: number; user: string; };
export type TDBOptions = TMVOptions & {
	insertOptions?: { onConflict: string; };
};
export type TDomainFields = { [key: string]: any; };
export type TExecutor = pg.Pool | pg.PoolClient | pg.Client;
export type TMVOptions = {
	isLoggerEnabled?: boolean;
	logger?: SharedTypes.TLogger;
	client?: TExecutor;
};
export type TSearchParams<T> =
	& TSearchParamsStrict<T>
	// & Partial<Record<JsonKeysToStringResult<Required<T>>, TDefault>>
	& { [key: string]: TDefault | undefined; };

type TSearchParamValue<T> = T extends object
	? TSearchParamObjectValue<T>
	: TSearchParamPrimitiveValue<T>;

type TSearchParamObjectValue<T> = T extends Date
	? BaseDate
	: T extends Buffer
		? BaseBuffer
		: T extends Array<unknown>
			? BaseArray<T>
			: BaseObject<T>;

type TSearchParamPrimitiveValue<T> = T extends string | number
	? BaseStringOrNumber<T>
	: BaseBoolean<ClearBoolean>;

type TDefault =
	| BaseDate
	| BaseBuffer
	| BaseStringOrNumber<ClearString>
	| BaseStringOrNumber<ClearNumber>
	| BaseBoolean<ClearBoolean>
	| BaseArray<any[]>
	| BaseObject<object>
	| null
	| { $eq: null; };
// ...
type BaseBuffer =
	| ClearBuffer
	| Base<ClearBuffer>
	| Array<Base<ClearBuffer>>;

type BaseDate =
	| ClearDate
	| Base<ClearDate>
	| Array<Base<ClearDate>>;

// https://www.postgresql.org/docs/current/ltree.html
type BaseLtree =
	| { ["$@>"]: string | string[]; }
	| { ["$<@"]: string | string[]; }
	| { ["$@"]: string | string[]; }
	| { ["$~"]: string | string[]; }
	| { ["$?"]: string | string[]; };

type BaseArray<T> =
	| T extends Date ? ClearDate : T
	| { ["$@>"]: NonNullable<T>; }
	| { ["$<@"]: NonNullable<T>; }
	| { ["$&&"]: NonNullable<T>; }
	| { $eq: T; }
	| { $ne: NonNullable<T> | null; };

// https://www.postgresql.org/docs/current/functions-json.html#FUNCTIONS-JSON-OP-TABLE
type BaseObject<T> =
	| { ["$?"]: string; }
	| { ["$@>"]: string | number | Date | string[] | number[] | Date[]; }
	| { $json: T; }
	| { $jsonb: T; }
	| { $eq: T; }
	| { $ne: NonNullable<T> | null; };

type BaseStringOrNumber<T> =
	| T
	| BaseLtree
	| Base<T>
	| Array<BaseLtree | Base<T>>;

type BaseBoolean<T> =
	| T
	| { $ne: NonNullable<T> | null; }
	| { $custom: { sign: string; value: string | number; }; };
// ...
type ClearString = Omit<string, Extract<keyof string, string>>;
type ClearNumber = Omit<number, Extract<keyof number, string>>;
type ClearBoolean = Omit<boolean, Extract<keyof boolean, string>>;

export type TSearchParamsStrict<T> = {
	[key in keyof T]: (null extends T[key] ? (null | { $eq: null; }) | TSearchParamValue<NonNullable<T[key]>> : TSearchParamValue<NonNullable<T[key]>>)
};

export type TTable<T extends readonly string[] | string[] = readonly string[]> = {
	additionalSortingFields?: string[];
	createField: { title: T[number]; type: "unix_timestamp" | "timestamp"; } | null;
	primaryKey: T[number] | [T[number], T[number], ...T[number][]] | null;
	tableFields: T;
	tableName: string;
	updateField: { title: T[number]; type: "unix_timestamp" | "timestamp"; } | null;
};